requirements.txt
asgiref==3.7.2
Django==4.2.3
django-appconf==1.0.5
django-cors-headers==4.2.0
django-nested-admin==4.0.2
django-staticfiles==1.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
python-monkey-business==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
